pyproject.toml
[project]
name = "sz-detection-challenge2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.1.3",
    "mne>=1.8.0",
    "polars>=1.17.1",
    "pytest>=8.3.4",
    "zarr>=2.18.4",
    # "epileptology",
    "pybids>=0.18.1",
    "networkx>=3.4.2",
    "python-louvain>=0.16",
    "numpy<2.0a0",
    "lightgbm>=4.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/szdetect"]
